Done, But No Validation

+++ pages/delete/[id].js
@@ -1,5 +1,16 @@
+import nookies from "nookies";
+
 export async function getServerSideProps(context) {
   const { id } = context.query;
+  const cookies = nookies.get(context);
+
+  if (!cookies.token) {
+    return {
+      redirect: {
+        destination: "/login",
+      },
+    };
+  }
 
   const req = await fetch(`${process.env.NEXT_PUBLIC_URL}/${id}`, {
     method: "DELETE",
